move ClientSideOcrd{Agent,File} to ocrd_models
