+++ ocrd/ocrd/mets_server.py
@@ -15,7 +15,7 @@ from pydantic import BaseModel, Field, ValidationError
 
 import uvicorn
 
-from ocrd_models import OcrdMets, OcrdFile, OcrdAgent
+from ocrd_models import OcrdMets, OcrdFile, ClientSideOcrdFile, OcrdAgent, ClientSideOcrdAgent
 from ocrd_utils import initLogging, getLogger, deprecated_alias
 
 #
@@ -79,74 +79,6 @@ class OcrdAgentListModel(BaseModel):
 #
 # Client
 #
-
-class ClientSideOcrdFile:
-    """
-    Provides the same interface as :py:class:`ocrd_models.ocrd_file.OcrdFile`
-    but without attachment to :py:class:`ocrd_models.ocrd_mets.OcrdMets` since
-    this represents the response of the :py:class:`ocrd.mets_server.OcrdMetsServer`.
-    """
-
-    def __init__(self, el, mimetype=None, pageId=None, loctype='OTHER', local_filename=None, mets=None, url=None, ID=None, fileGrp=None):
-        """
-        Args:
-            el (): ignored
-        Keyword Args:
-            mets (): ignored
-            mimetype (string): ``@MIMETYPE`` of this ``mets:file``
-            pageId (string): ``@ID`` of the physical ``mets:structMap`` entry corresponding to this ``mets:file``
-            loctype (string): ``@LOCTYPE`` of this ``mets:file``
-            url (string): ignored XXX the remote/original file once we have proper mets:FLocat bookkeeping 
-            local_filename (): ``@xlink:href`` of this ``mets:file`` - XXX the local file once we have proper mets:FLocat bookkeeping
-            ID (string): ``@ID`` of this ``mets:file``
-        """
-        self.ID = ID
-        self.mimetype = mimetype
-        self.local_filename = local_filename
-        self.loctype = loctype
-        self.pageId = pageId
-        self.fileGrp = fileGrp
-
-    def __str__(self):
-        props = ', '.join([
-            '='.join([k, getattr(self, k) if hasattr(self, k) and getattr(self, k) else '---'])
-            for k in ['fileGrp', 'ID', 'mimetype', 'url', 'local_filename']
-        ])
-        return '<OcrdFile %s]/>' % (props)
-
-class ClientSideOcrdAgent():
-    """
-    Provides the same interface as :py:class:`ocrd_models.ocrd_file.OcrdAgent`
-    but without attachment to :py:class:`ocrd_models.ocrd_mets.OcrdMets` since
-    this represents the response of the :py:class:`ocrd.mets_server.OcrdMetsServer`.
-    """
-
-    def __init__(self, el, name=None, _type=None, othertype=None, role=None, otherrole=None,
-                 notes=None):
-        """
-        Args:
-            el (): ignored
-        Keyword Args:
-            name (string):
-            _type (string):
-            othertype (string):
-            role (string):
-            otherrole (string):
-            notes (dict):
-        """
-        self.name = name
-        self.type = _type
-        self.othertype = othertype
-        self.role = role
-        self.otherrole = otherrole
-        self.notes = notes
-
-    def __str__(self):
-        props = ', '.join([
-            '='.join([k, getattr(self, k) if getattr(self, k) else '---'])
-            for k in ['type', 'othertype', 'role', 'otherrole', 'name']
-        ])
-        return '<OcrdAgent [' + props + ']/>'
 
 
 class ClientSideOcrdMets():
