+++ ocrd_models/ocrd_models/__init__.py
@@ -1,9 +1,9 @@
 """
 APIs and schemas for various file formats in the OCR domain.
 """
-from .ocrd_agent import OcrdAgent
+from .ocrd_agent import OcrdAgent, ClientSideOcrdAgent
 from .ocrd_exif import OcrdExif
-from .ocrd_file import OcrdFile
+from .ocrd_file import OcrdFile, ClientSideOcrdFile
 from .ocrd_mets import OcrdMets
 from .ocrd_xml_base import OcrdXmlDocument
 from .report import ValidationReport
