+++ ocrd_models/ocrd_models/ocrd_file.py
@@ -227,3 +227,39 @@ class OcrdFile():
         if el_FLocat is None:
             el_FLocat = ET.SubElement(self._el, TAG_METS_FLOCAT)
         el_FLocat.set("{%s}href" % NS["xlink"], url)
+
+
+class ClientSideOcrdFile:
+    """
+    Provides the same interface as :py:class:`ocrd_models.ocrd_file.OcrdFile`
+    but without attachment to :py:class:`ocrd_models.ocrd_mets.OcrdMets` since
+    this represents the response of the :py:class:`ocrd.mets_server.OcrdMetsServer`.
+    """
+
+    def __init__(self, el, mimetype=None, pageId=None, loctype='OTHER', local_filename=None, mets=None, url=None, ID=None, fileGrp=None):
+        """
+        Args:
+            el (): ignored
+        Keyword Args:
+            mets (): ignored
+            mimetype (string): ``@MIMETYPE`` of this ``mets:file``
+            pageId (string): ``@ID`` of the physical ``mets:structMap`` entry corresponding to this ``mets:file``
+            loctype (string): ``@LOCTYPE`` of this ``mets:file``
+            url (string): ignored XXX the remote/original file once we have proper mets:FLocat bookkeeping 
+            local_filename (): ``@xlink:href`` of this ``mets:file`` - XXX the local file once we have proper mets:FLocat bookkeeping
+            ID (string): ``@ID`` of this ``mets:file``
+        """
+        self.ID = ID
+        self.mimetype = mimetype
+        self.local_filename = local_filename
+        self.url = url
+        self.loctype = loctype
+        self.pageId = pageId
+        self.fileGrp = fileGrp
+
+    def __str__(self):
+        props = ', '.join([
+            '='.join([k, getattr(self, k) if hasattr(self, k) and getattr(self, k) else '---'])
+            for k in ['fileGrp', 'ID', 'mimetype', 'url', 'local_filename']
+        ])
+        return '<ClientSideOcrdFile %s]/>' % (props)
